# perf: reduce networkbehaviourilpp type resolution time